+++ src/components/FormView.jsx
@@ -1,17 +1,15 @@
 import React from 'react';
 import { useIntl, defineMessages } from 'react-intl';
-import {
-  Segment,
-  Message,
-  Grid,
-  Form,
-  Progress,
-  Button,
-} from 'semantic-ui-react';
+import { Segment, Message, Grid, Form, Progress } from 'semantic-ui-react';
 import { getFieldName } from 'volto-form-block/components/utils';
 import Field from 'volto-form-block/components/Field';
+import {
+  OTPWidget,
+  OTP_FIELDNAME_EXTENDER,
+  Button,
+} from 'volto-form-block/components/Widget';
 import config from '@plone/volto/registry';
-
+import { FormResult } from 'volto-form-block/components';
 /* Style */
 import 'volto-form-block/components/FormView.css';
 
@@ -28,17 +26,9 @@ const messages = defineMessages({
     id: 'Error',
     defaultMessage: 'Error',
   },
-  success: {
-    id: 'form_submit_success',
-    defaultMessage: 'Sent!',
-  },
-  empty_values: {
-    id: 'form_empty_values_validation',
-    defaultMessage: 'Fill in the required fields',
-  },
-  reset: {
-    id: 'form_reset',
-    defaultMessage: 'Clear',
+  form_errors: {
+    id: 'form_errors_validation',
+    defaultMessage: 'There are some errors in the form.',
   },
 });
 
@@ -53,31 +43,41 @@ const FormView = ({
   resetFormOnError,
   captcha,
   id,
+  getErrorMessage,
+  path,
+  block_id,
 }) => {
   const intl = useIntl();
   const FieldSchema = config.blocks.blocksConfig.form.fieldSchema;
 
   const isValidField = (field) => {
-    return formErrors?.indexOf(field) < 0;
+    return formErrors?.filter((e) => e.field === field).length === 0;
   };
 
-  /* Function that replaces variables from the user customized message  */
-  const replaceMessage = (text) => {
-    let i = 0;
-    while (i < data.subblocks.length) {
-      let idField = getFieldName(
-        data.subblocks[i].label,
-        data.subblocks[i].field_id,
-      );
-      text = text.replaceAll(
-        '${' + idField + '}',
-        formData[idField]?.value || '',
-      );
-      i++;
+  const submit = (e) => {
+    resetFormOnError();
+    onSubmit(e);
+  };
+
+  const getFieldsToSendWithValue = (subblock) => {
+    var fields_to_send = [];
+    var fieldSchemaProperties = FieldSchema(subblock)?.properties;
+    for (var key in fieldSchemaProperties) {
+      if (fieldSchemaProperties[key].send_to_backend) {
+        fields_to_send.push(key);
+      }
     }
-    return text;
-  };
 
+    var fields_to_send_with_value = Object.assign(
+      {},
+      ...fields_to_send.map((field) => {
+        return {
+          [field]: subblock[field],
+        };
+      }),
+    );
+    return fields_to_send_with_value;
+  };
   return (
     <div className="block form">
       <div className="public-ui">
@@ -86,44 +86,18 @@ const FormView = ({
           {data.description && (
             <p className="description">{data.description}</p>
           )}
-          {formState.error ? (
-            <Message error role="alert">
-              <Message.Header as="h4">
-                {intl.formatMessage(messages.error)}
-              </Message.Header>
-              <p>{formState.error}</p>
-              <Button secondary type="clear" onClick={resetFormOnError}>
-                {intl.formatMessage(messages.reset)}
-              </Button>
-            </Message>
-          ) : formState.result ? (
-            <Message positive role="alert">
-              {/* Custom message */}
-              {data.send_message ? (
-                <p
-                  dangerouslySetInnerHTML={{
-                    __html: replaceMessage(data.send_message),
-                  }}
-                />
-              ) : (
-                <>
-                  {/* Default message */}
-                  <Message.Header as="h4">
-                    {intl.formatMessage(messages.success)}
-                  </Message.Header>
-                  <p>{formState.result}</p>
-                </>
-              )}
-              {/* Back button */}
-              <Button secondary type="clear" onClick={resetFormState}>
-                {intl.formatMessage(messages.reset)}
-              </Button>
-            </Message>
+
+          {formState.result ? (
+            <FormResult
+              formState={formState}
+              data={data}
+              resetFormState={resetFormState}
+            />
           ) : (
             <Form
               id={id}
               loading={formState.loading}
-              onSubmit={onSubmit}
+              onSubmit={submit}
               autoComplete="off"
               method="post"
             >
@@ -151,22 +125,8 @@ const FormView = ({
                 {data.subblocks?.map((subblock, index) => {
                   let name = getFieldName(subblock.label, subblock.id);
 
-                  var fields_to_send = [];
-                  var fieldSchemaProperties = FieldSchema(subblock)?.properties;
-                  for (var key in fieldSchemaProperties) {
-                    if (fieldSchemaProperties[key].send_to_backend) {
-                      fields_to_send.push(key);
-                    }
-                  }
-
-                  var fields_to_send_with_value = Object.assign(
-                    {},
-                    ...fields_to_send.map((field) => {
-                      return {
-                        [field]: subblock[field],
-                      };
-                    }),
-                  );
+                  const fields_to_send_with_value =
+                    getFieldsToSendWithValue(subblock);
 
                   return (
                     <Grid.Row key={'row' + index}>
@@ -188,25 +148,90 @@ const FormView = ({
                               : formData[name]?.value
                           }
                           valid={isValidField(name)}
+                          errorMessage={getErrorMessage(name)}
                           formHasErrors={formErrors?.length > 0}
                         />
                       </Grid.Column>
                     </Grid.Row>
                   );
                 })}
+
+                {/*OTP*/}
+                {data.email_otp_verification ? (
+                  data.subblocks
+                    .filter((subblock) => subblock.use_as_bcc)
+                    .map((subblock, index) => {
+                      const fieldName = getFieldName(
+                        subblock.label,
+                        subblock.id,
+                      );
+                      const name = fieldName + OTP_FIELDNAME_EXTENDER;
+                      const fieldValue = formData[fieldName]?.value;
+                      const value = formData[fieldName]?.otp;
+                      const fields_to_send_with_value =
+                        getFieldsToSendWithValue(subblock);
+
+                      return (
+                        <Grid.Row key={'row_otp' + index}>
+                          <Grid.Column>
+                            <OTPWidget
+                              {...subblock}
+                              fieldValue={fieldValue}
+                              onChange={(field, value) =>
+                                onChangeFormData(
+                                  subblock.id,
+                                  fieldName,
+                                  fieldValue,
+                                  {
+                                    ...fields_to_send_with_value,
+                                    otp: value,
+                                  },
+                                )
+                              }
+                              value={value}
+                              valid={isValidField(name)}
+                              errorMessage={getErrorMessage(name)}
+                              formHasErrors={formErrors?.length > 0}
+                              path={path}
+                              block_id={block_id}
+                            />
+                          </Grid.Column>
+                        </Grid.Row>
+                      );
+                    })
+                ) : (
+                  <></>
+                )}
+
                 {captcha.render()}
                 {formErrors.length > 0 && (
                   <Message error role="alert">
                     <Message.Header as="h4">
                       {intl.formatMessage(messages.error)}
                     </Message.Header>
-                    <p>{intl.formatMessage(messages.empty_values)}</p>
+                    <p>{intl.formatMessage(messages.form_errors)}</p>
+                  </Message>
+                )}
+                {formState.error && (
+                  <Message error role="alert">
+                    <Message.Header as="h4">
+                      {intl.formatMessage(messages.error)}
+                    </Message.Header>
+                    <p>{formState.error}</p>
                   </Message>
                 )}
                 <Grid.Row centered className="row-padded-top">
                   <Grid.Column textAlign="center">
                     {data?.show_cancel && (
-                      <Button secondary type="clear" onClick={resetFormState}>
+                      <Button
+                        secondary
+                        type="button"
+                        onClick={(e) => {
+                          e.preventDefault();
+                          e.stopPropagation();
+                          resetFormState();
+                        }}
+                      >
                         {data.cancel_label ||
                           intl.formatMessage(messages.default_cancel_label)}
                       </Button>