+++ src/components/Widget/TextareaWidget.jsx
@@ -10,7 +10,7 @@ import PropTypes from 'prop-types';
 import { Label, TextArea } from 'semantic-ui-react';
 
 import { injectIntl } from 'react-intl';
-import { FormFieldWrapper } from '@plone/volto/components';
+import FormFieldWrapper from '@plone/volto/components/manage/Widgets/FormFieldWrapper';
 
 /**
  * TextareaWidget, a widget for multiple lines text
@@ -25,15 +25,8 @@ import { FormFieldWrapper } from '@plone/volto/components';
  * ```
  */
 const TextareaWidget = (props) => {
-  const {
-    id,
-    maxLength,
-    value,
-    onChange,
-    placeholder,
-    required,
-    invalid,
-  } = props;
+  const { id, maxLength, value, onChange, placeholder, required, invalid } =
+    props;
   const [lengthError, setlengthError] = useState('');
 
   const onhandleChange = (id, value) => {